two more lesson

+++ permMissingElement/src/permMissingElement/Solution.java
@@ -7,12 +7,12 @@ public class Solution {
         		// instead of two for loops
 		// seq numbers sum rule = (A.length + 1 +1) * A.length /2
 		
-		int sum = (A.length+1 +1) * (A.length+1) /2;
+		long sum = (A.length+1 +1) * (A.length+1) /2;
 		for (int i=0; i<A.length ; i++) {
 			sum = sum - A[i];
 		}
 		
-		return sum;
+		return (int)sum;
     }
 
 }
